refactor(stackflow): split transition-animation into dom.ts + animation.ts
- Add data-part attributes to AppBarMain, AppBarIconButton, AppBarSlot
  to replace fragile className-based selectors
- Split transition-animation.ts into:
  - dom.ts: DOM discovery (findTransitionTargets) and inline style management
  - animation.ts: WAAPI animation functions
- Unify SwipeEndpoints with IOS_ONSCREEN/IOS_OFFSCREEN constants
- Remove transition-animation.ts
